reorganized some components and updated import paths
