+++ src/components/auth/LoginForm.tsx
@@ -7,7 +7,7 @@ import { zodResolver } from "@hookform/resolvers/zod"
 import { useRouter } from "next/navigation"
 import { useState } from "react"
 import { useForm } from "react-hook-form"
-import Button from "../base/Button"
+import Button from "../base/button/Button"
 import Input from "../base/input/Input"
 import Link from "../base/Link"
 import toast from "../base/toast/Toast"
