+++ src/components/auth/RegistrationForm.tsx
@@ -5,7 +5,7 @@ import { createUser } from "@/services/auth"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { useRouter } from "next/navigation"
 import { useForm } from "react-hook-form"
-import Button from "../base/Button"
+import Button from "../base/button/Button"
 import Input from "../base/input/Input"
 import Link from "../base/Link"
 import toast from "../base/toast/Toast"
