+++ src/components/home/skeletons/EventSliderSkeleton.tsx
