+++ src/lib/payload.ts
@@ -1,4 +1,6 @@
 import configPromise from '@payload-config'
 import { getPayload } from 'payload'
 
-export const payloadClient = await getPayload({ config: configPromise })
+export const payloadClient = await getPayload({ 
+    config: configPromise 
+})
